+++ miniRT/v1/geometry/intersection.c
@@ -1,10 +1,10 @@
 #include "libgeometry.h"
 #include <math.h>
 
-float		inter_pl_r(t_plane pl, t_ray r)
+double		inter_pl_r(t_plane pl, t_ray r)
 {
-	float t;
-	float d;
+	double t;
+	double d;
 
 	if (dot_product(pl.vec, r.vec) < 1e-5)
 	{
@@ -15,7 +15,7 @@ float		inter_pl_r(t_plane pl, t_ray r)
 	return (t);
 }
 
-t_point		point_plus_vec(t_point po, float t, t_vector vec)
+t_point		point_plus_vec(t_point po, double t, t_vector vec)
 {
 	return point(po.x + t * vec.x, po.y + t * vec.y, po.z + t * vec.z);
 }
